# Faster and more readable implementation of Hash#deep_merge