@rafaelfranca 2017-08-15
Can you use benchmark-ips in your benchmark?