@msimonborg 2017-08-15
Fixed that LOC per your comment and changed benchmarks to `benchmark/ips`. Let me know if there's anything else you'd want to see!