@jonathanhefner 2017-08-16
I ran the benchmark on my machine (Ruby 2.2), and here are the numbers I got:
```
new_deep_merge_all_dupes:    99330.4 i/s
old_deep_merge_all_dupes:    94990.2 i/s - 1.05x  slower

new_deep_merge_some_dupes:   105237.5 i/s
old_deep_merge_some_dupes:    91174.9 i/s - 1.15x  slower

new_deep_merge_no_dupes:    95207.6 i/s
old_deep_merge_no_dupes:    84227.6 i/s - same-ish: difference falls within error
```

Which is still an improvement. 👍 

Also, a while back, I wrote a gem ([benchmark-inputs](https://github.com/jonathanhefner/benchmark-inputs)) to help with benchmarking multiple use cases, particularly when mutation is involved.  Using it to benchmark the bang method, here are the numbers I got:

```
NEW deep_merge!:    183292.1 i/s
OLD deep_merge!:    152630.4 i/s - 1.20x slower
```

And here's the script to reproduce that:
```ruby
require 'active_support'
require 'benchmark/inputs'

class Hash
  def new_deep_merge(other_hash, &block)
    dup.new_deep_merge!(other_hash, &block)
  end

  def new_deep_merge!(other, &block)
    merge!(other) do |key, old_val, new_val|
      if old_val.is_a?(Hash) && new_val.is_a?(Hash)
        old_val.new_deep_merge(new_val, &block)
      elsif block_given?
        block.call(key, old_val, new_val)
      else
        new_val
      end
    end
  end
end

HASHES = [
  { a: false, b: { c: [3, 4, 5] }, d: 2 },
  { a: false, b: { x: [3, 4, 5] }, e: 2, f: { y: false } },
  { e: 2, f: { y: false }, g: 'hi', h: true },
]

OTHER = { a: true, b: { c: [1, 2, 3] }, d: 1 }

Benchmark.inputs(HASHES) do |job|
  job.dup_inputs = true
  job.report('OLD deep_merge!'){|h| h.deep_merge!(OTHER) }
  job.report('NEW deep_merge!'){|h| h.new_deep_merge!(OTHER) }
  job.compare!
end
```