@msimonborg 2017-08-16
Rebased and squashed. Thanks @lugray !